move prompts from environ.py to prompt/
